PLAYA

Documentation

To add Package use:

julia> Pkg.add("https://github.com/tungli/PLAYA.jl.git")

A try to replicate a study of air kinetics from [1] is in demo/air-kinetics.jl.

PLAYA.parse_reactions — Function

parse_reactions(filename, env; to_replace=[])

Parse reactions from a file and output a Catalyst ReactionSystem and the rate values.

Inputs

File name.

Environment of variables that get substituted into rates.

Optional replacement list (Array{Pair{String, String}}) that translates the

species names (mainly to replace symbols that Catalyst complains about).

Reaction file

On each line there are two seperators: ':' and '!' and three data fields. At the end of the line a comment starting with '#' may be inserted. Each line is one (forward) reaction:

id: reaction ! rate expression (# optional comment)

The reaction is specified same as a reaction in Catalyst but without the "rate part". This imposes limitations on the use of parenthesis and symbols that represent functions in Julia.

Rate expression can be one either a function or a number. Evaluations are done using the env::Dict which acts as an environment for variables.

Additionally, to avoid long functions, a "subdict" with functions can be used with the following syntax:

subdict { key }

which evaluates to:

env[Symbol(subdict)](key::String, env)

For example:

BOLOS { A -> B } * 1e6

will search for the BOLOS dict under the keyword "A -> B", evaluate the found function and substitute the corresponding value and evaluate the multiplication.

PLAYA.safe_replace — Function

safe_replace(s, l)

Replaces multiple strings "safely" – longer strings are replace first, and a string cannot be replaced twice.

PLAYA.parse_bolos_output — Function

Parses one block of BOLOS output file. Return a Dict{String, Interpolation} where the String is the reaction.

PLAYA.keep_fixed! — Function

keep_fixed!(odesys::ModelingToolkit.AbstractODESystem, species_name)

Modify the system of ODEs, such that the concentration of species will be fixed, i.e. its time derivative will be zero.

Inputs

odesys - system of ODEs (from DifferentialEquations.jl)

species_name - array of species identifiers

1A. Obrusník, P. Bílek, T. Hoder, M. Šimek, Z. Bonaventura, Plasma Sources Sci. Technol., 2018, DOI:10.1088/1361-6595/aad663.